pyproject.toml
[project]
name = "uidai_py"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Arun Khattri", email = "arunkhattri@outlook.com"},
]
dependencies = [
    "lxml<5,>=4.2.1",
    "pillow>=10.2.0",
    "cryptography>=42.0.5",
    "xmlsec>=1.3.13",
    "signxml==3.0.1",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "GPL-3-or-later"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "black>=24.2.0",
    "isort>=5.13.2",
    "mypy>=1.8.0",
    "python-lsp-server>=1.10.0",
    "python-lsp-ruff>=2.2.0",
    "pylint>=3.1.0",
    "pytest>=8.1.0",
    "ipython>=8.22.1",
    "python-decouple>=3.8",
]
